refactor: rename orgProfile translations to organizations
