+++ src/core/translation/locales/jp/organizations/organizations.json
@@ -8,5 +8,6 @@
   "org-profile-description": "公開説明",
   "org-profile-description-placeholder": "説明を入力してください...",
   "org-profile-optional": "任意",
-  "org-profile-save-button": "変更を保存"
+  "org-profile-save-button": "変更を保存",
+  "org-profile-create-button": "組織を作る"
 }
